Implement texture loading on a key
Currently loads doubly due to sticky keys so we'll need a way
around this. The idea is to keep a second name that checks the
first. This will make the double-loading problem easy.
We also need to deallocate the previous buffer.

It does take a considerable amount of time to load though...